fix: resolve all 856 lint issues from 12 new linters
- err113 (381): sentinel errors in 19 new errors.go files across all packages
- goimports (145): auto-formatted all import groups
- sloglint (2): remaining context issues
- gosec (58): file permissions 0o644→0o600, ReadHeaderTimeout, safe uint32 casts
- misspell (37): cancelled→canceled, serialises→serializes, etc.
- unparam (35): removed/prefixed unused params, simplified signatures
- gocyclo (17): nolint for genuine state machines/parsers
- revive (10): error naming, context-as-first-arg, blank import comments
- dupl (8): extracted shared helpers (runWorkers, checkCalibFields, resolveRunFile)
- thelper (9): added t.Helper() to test helpers
- unconvert (6): removed unnecessary type conversions
- funlen (8): nolint for complex orchestrators
- bodyclose (1), goconst (1), nolintlint (10)

29 linters, 0 issues. Pre-commit hook enforces lint-new on every commit.

+++ engine/batch_walk_test.go
@@ -84,7 +84,7 @@ func TestBatchWalk_Parallel(t *testing.T) {
 				cur := concurrency.Add(1)
 				for {
 					old := maxConcurrency.Load()
-					if int32(cur) <= old || maxConcurrency.CompareAndSwap(old, int32(cur)) {
+					if cur <= old || maxConcurrency.CompareAndSwap(old, cur) {
 						break
 					}
 				}
